+++ cmd/loupedeck/main.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"os"
+
 	"github.com/go-go-golems/glazed/pkg/cmds/logging"
 	"github.com/go-go-golems/glazed/pkg/help"
 	helpcmd "github.com/go-go-golems/glazed/pkg/help/cmd"
@@ -15,6 +17,9 @@ import (
 var version = "dev"
 
 func main() {
+	bootstrap, err := verbscmd.DiscoverBootstrap(os.Args[1:])
+	cobra.CheckErr(err)
+
 	rootCmd := &cobra.Command{
 		Use:     "loupedeck",
 		Short:   "Run Loupedeck Live JavaScript scenes and hardware workflows",
@@ -34,8 +39,11 @@ func main() {
 	cobra.CheckErr(err)
 	runCobraCmd, err := loupedeckcmdcommon.BuildCobraCommandDualMode(runCommand)
 	cobra.CheckErr(err)
+	rootCmd.PersistentFlags().StringArray(verbscmd.VerbRepositoryFlag, nil, "Additional JavaScript verb repositories to scan before command registration (repeatable)")
 	rootCmd.AddCommand(runCobraCmd)
-	rootCmd.AddCommand(verbscmd.NewCommand())
+	verbsCommand, err := verbscmd.NewCommand(bootstrap)
+	cobra.CheckErr(err)
+	rootCmd.AddCommand(verbsCommand)
 	rootCmd.AddCommand(doccmd.NewCommand())
 
 	cobra.CheckErr(rootCmd.Execute())